+++ edu/brown/securities/prediction/structures/PMNo.java
@@ -9,7 +9,7 @@ import brown.securities.prediction.PredictionMarket;
 
 /**
  * Private backend prediction market implementation
- * Agents are provided pointers to the public face
+ * Agents are provided podoubleers to the public face
  * to prevent illegal modifications
  * @author lcamery
  *
@@ -32,37 +32,37 @@ public class PMNo implements Security {
 	 * Invokes backend cost
 	 */
 	@Override
-	public double cost(int newq1, int newq2) {
+	public double cost(double newq1, double newq2) {
 		return this.backend.cost(newq1, newq2);
 	}
 	
 	/*
 	 * Quotes the cumulative price for a certain number of yes shares
-	 * @param shareNum : int
+	 * @param shareNum : double
 	 * @return cost : double
 	 */
 	@Override
-	public double bid(int shareNum) {
+	public double bid(double shareNum) {
 		return cost(0, shareNum);
 	}
 	
 	/*
 	 * Quotes the cumulative price for a certain number of no shares
-	 * @param shareNum : int
+	 * @param shareNum : double
 	 * @return cost : double
 	 */
 	@Override
-	public double ask(int shareNum) {
+	public double ask(double shareNum) {
 		return cost(0, -1 * shareNum);
 	}
 	
 	/*
 	 * Returns a transaction to an agent that buys yes
 	 * @param agentID : agent's public ID
-	 * @param shareNum : int
+	 * @param shareNum : double
 	 * @return share : share object; extendable in real games
 	 */
-	public Transaction buy(Integer agentID, int shareNum) {
+	public Transaction buy(Integer agentID, double shareNum) {
 		Transaction trans = new Transaction(this, shareNum, agentID, cost(0, shareNum));
 		backend.no(agentID, shareNum);
 		return trans;
@@ -71,10 +71,10 @@ public class PMNo implements Security {
 	/*
 	 * Returns a transaction to an agent that buys no
 	 * @param agentID : agent's public ID
-	 * @param shareNum : int
+	 * @param shareNum : double
 	 * @return share : share object; extendable in real games
 	 */
-	public Transaction sell(Integer agentID, int shareNum) {
+	public Transaction sell(Integer agentID, double shareNum) {
 		Transaction trans = new Transaction(this, shareNum, agentID, cost(0, -1 * shareNum));
 		backend.no(agentID, -1*shareNum);
 		return trans;
